Correctly categorize test as a unit one
