+++ services/ingestion/cleaner/test/unit/events/onObjectRemoved.spec.js
@@ -1,3 +1,4 @@
+import AWS from 'aws-sdk-mock';
 import { promisify } from 'util';
 import onObjectRemoved from '../../../src/events/onObjectRemoved';
 
@@ -38,14 +39,70 @@ describe(`Function onObjectRemoved in "@eubfr/ingestion-cleaner"`, () => {
 
   test('The function expects a message body from the actual event stub', () => {
     process.env.BUCKET = 'foo';
-    // Remove the message from the actual event to check that the function asserts it.
-    delete eventEubfr.Records[0].Sns.Message;
+    // The JSON.stringify is used as a simple means of cloning the original object.
+    const copy = JSON.parse(JSON.stringify(eventEubfr));
 
+    // Remove the message from the actual event to check that the function asserts it.
+    delete copy.Records[0].Sns.Message;
+
+    const context = {};
+
+    expect.assertions(1);
+    const result = handler(copy, context);
+    return expect(result).rejects.toBe(`Missing message body`);
+  });
+});
+
+describe(`Function onObjectRemoved in "@eubfr/ingestion-cleaner" when S3 service works well`, () => {
+  beforeAll(() => {
+    AWS.mock('S3', 'deleteObject', (params, callback) => {
+      callback(null, 'Successfully deleted a file from the S3 storage');
+    });
+  });
+
+  afterEach(() => {
+    // ensure that the environment is not left over a test
+    delete process.env.BUCKET;
+  });
+
+  afterAll(() => {
+    AWS.restore('S3', 'deleteObject');
+  });
+
+  test('The function removes an object from S3 with the actual event stub', () => {
+    process.env.BUCKET = 'foo';
+    const event = eventEubfr;
+    const context = {};
+
+    const result = handler(event, context);
+    expect.assertions(1);
+    return expect(result).resolves.toBe(`object removed`);
+  });
+});
+
+describe(`Function onObjectRemoved in "@eubfr/ingestion-cleaner" when S3 service fails`, () => {
+  beforeAll(() => {
+    AWS.mock('S3', 'deleteObject', (params, callback) => {
+      callback('S3 failure');
+    });
+  });
+
+  afterEach(() => {
+    // ensure that the environment is not left over a test
+    delete process.env.BUCKET;
+  });
+
+  afterAll(() => {
+    AWS.restore('S3', 'deleteObject');
+  });
+
+  test('The function fails to remove an object from S3 because of service failure', () => {
+    process.env.BUCKET = 'foo';
     const event = eventEubfr;
     const context = {};
 
     expect.assertions(1);
     const result = handler(event, context);
-    return expect(result).rejects.toBe(`Missing message body`);
+    return expect(result).rejects.toBe(`S3 failure`);
   });
 });
